feat: Add quest Platinum Cafe member (#48)
* Platinum Cafe member quest: add quest
* Platinum Cafe Customer quest: restrict to only spawning in acts 1 or 2
* Platinum Cafe Member quest: add missing requiredModId to patches

Other than all the reflection, the one noteworthy thing is that I added a show method to the quest tracker system. I didn't see any existing way to make hidden trackers visible and there weren't any examples using it, so my assumption is that this part of the tracker system that wasn't fully completed.