feat: Add quest BountyIC (#52)
This quest adds a new monster, some vfx and 3 cards taken from StS2 teasers. When the quest is taken, the next elite this act is replaced by The Ironclad who has a simple attack pattern that starts with Demon Form then goes over to 3 cycling attacks. As a reward, the relic is replaced with Burning blood and the card reward with custom IC cards.

I'll leave some comments behind for things I want to highlight for review, would also appreciate feedback on monster balancing. Not sure I've hit a sweet spot there yet.

* feat: Add card Unrelenting

* chore: Add image for card Unrelenting

* feat: Add card Taunt

* feat: Add card StoneArmor

* feat: Buff Taunt upgrade by 1 Block

* feat: Add quest BountyICQuest

TODO: Add new Monster Ironclad and replace the next elite with him

* feat: Add Ironclad monster

Includes logic to replace the elite and and vfx to sell the idea

* fix: Fix replacing elites when the quest has already failed

* chore: Remove some unused comments

* feat: Add small character movement animations to ICmonster moves

* feat: Visual updates to FakePlayCardEffect

- Increase travel curve by 20px
- Fade card out during travel
- Reduce action duration to 1.75f

* fix: Fix Encounter not showing up in run history

* feat: Give Ironclad the Fossilized helix relic

* fix: Fix trigger condition of BountyIC to respect other elite replacements

* fix: Change reward generation logic to better respect upgrade logic